table_process/merge_chains/requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
python-Levenshtein>=0.12.2

# For embeddings (optional but recommended)
torch>=1.9.0
transformers>=4.20.0
scikit-learn>=0.24.0
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.62.0
